fix migrations

+++ database/migrations/2016_08_18_063337_create_admins_table.php
@@ -7,11 +7,13 @@ class CreateAdminsTable extends Migration
 {
     /**
      * Run the migrations.
-     *
      * @return void
      */
     public function up()
     {
+        if (Schema::hasTable('admins')) {
+            return;
+        }
         Schema::create('admins', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
@@ -21,21 +23,23 @@ class CreateAdminsTable extends Migration
             $table->string('admin_rand', 32)->comment('随机加密值');
             $table->mediumText('group_id')->nullable()->comment('NULL不属于任何组');
             $table->mediumText('privilege')->nullable()->comment('all是全部权限');
-            $table->timestamp('last_time')->comment('活跃时间');
+            $table->timestamp('last_time')->nullable()->comment('活跃时间');
             $table->ipAddress('login_ip')->comment('活跃IP');
             $table->tinyInteger('login_num')->comment('尝试登录数');
-            $table->timestamp('lock_time')->comment('登录锁定时间');
+            $table->timestamp('lock_time')->nullable()->comment('登录锁定时间');
             $table->tinyInteger('is_enable')->comment('是否启用');
         });
     }
 
     /**
      * Reverse the migrations.
-     *
      * @return void
      */
     public function down()
     {
+        if (!Schema::hasTable('admins')) {
+            return;
+        }
         Schema::drop('admins');
     }
 }
