+++ database/migrations/2016_08_18_063341_create_assesses_table.php
@@ -12,6 +12,9 @@ class CreateAssessesTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('assesses')) {
+            return;
+        }
         Schema::create('assesses', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
@@ -19,8 +22,8 @@ class CreateAssessesTable extends Migration
             $table->string('title', 64)->comment('考核表名称');
             $table->string('explains', 256)->comment('考核说明');
             $table->tinyInteger('group_level')->comment('定义考核分组级别');
-            $table->integer('start_time')->unsigned()->comment('开始时间');
-            $table->integer('end_time')->unsigned()->comment('结束时间');
+            $table->timestamp('start_time')->nullable()->comment('开始时间');
+            $table->timestamp('end_time')->nullable()->comment('结束时间');
             $table->tinyInteger('is_enable')->comment('是否启用');
             $table->string('target', 64)->comment('考核目标');
             $table->mediumText('ext_info')->nullable()->comment('条目数组');
@@ -34,6 +37,9 @@ class CreateAssessesTable extends Migration
      */
     public function down()
     {
+        if (!Schema::hasTable('assesses')) {
+            return;
+        }
         Schema::drop('assesses');
     }
 }
