+++ database/migrations/2016_08_18_063342_create_assess_logs_table.php
@@ -12,6 +12,9 @@ class CreateAssessLogsTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('assess_logs')) {
+            return;
+        }
         Schema::create('assess_logs', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
@@ -19,7 +22,7 @@ class CreateAssessLogsTable extends Migration
             $table->integer('assess_id')->unsigned()->comment('考核表编号');
             $table->integer('grade_id')->unsigned()->comment('评分人');
             $table->integer('re_grade_id')->unsigned()->comment('被评分人');
-            $table->timestamp('add_time')->comment('添加时间');
+            $table->timestamp('add_time')->nullable()->comment('添加时间');
             $table->mediumText('score')->nullable()->comment('评分数组');
         });
     }
@@ -31,6 +34,9 @@ class CreateAssessLogsTable extends Migration
      */
     public function down()
     {
+        if (!Schema::hasTable('assess_logs')) {
+            return;
+        }
         Schema::drop('assess_logs');
     }
 }
