+++ database/migrations/2016_08_18_063343_create_itlinks_table.php
@@ -12,6 +12,9 @@ class CreateItlinksTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('itlinks')) {
+            return;
+        }
         Schema::create('itlinks', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
@@ -19,8 +22,8 @@ class CreateItlinksTable extends Migration
             $table->string('name', 128)->comment('名称');
             $table->string('short_name', 32)->comment('调用短名');
             $table->tinyInteger('is_enable')->comment('是否启用');
-            $table->timestamp('start_time')->comment('开始时间');
-            $table->timestamp('end_time')->comment('结束时间');
+            $table->timestamp('start_time')->nullable()->comment('开始时间');
+            $table->timestamp('end_time')->nullable()->comment('结束时间');
             $table->integer('max_show_num')->unsigned()->comment('最大显示数量');
             $table->integer('show_num')->unsigned()->comment('显示数量');
             $table->integer('max_hit_num')->unsigned()->comment('最大点击数');
@@ -36,6 +39,9 @@ class CreateItlinksTable extends Migration
      */
     public function down()
     {
+        if (!Schema::hasTable('itlinks')) {
+            return;
+        }
         Schema::drop('itlinks');
     }
 }
