+++ database/migrations/2016_08_18_063344_create_members_table.php
@@ -12,6 +12,9 @@ class CreateMembersTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('members')) {
+            return;
+        }
         Schema::create('members', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
@@ -20,11 +23,11 @@ class CreateMembersTable extends Migration
             $table->string('member_pwd', 32)->comment('密码');
             $table->string('member_rand', 32)->comment('随机加密值');
             $table->mediumText('group_id')->nullable()->comment('NULL不属于任何组');
-            $table->timestamp('register_time')->comment('注册时间');
-            $table->timestamp('last_time')->comment('活跃时间');
-            $table->ipAddress('login_ip')->comment('活跃IP');
+            $table->timestamp('register_time')->nullable()->comment('注册时间');
+            $table->timestamp('last_time')->nullable()->comment('活跃时间');
+            $table->ipAddress('login_ip')->nullable()->comment('活跃IP');
             $table->tinyInteger('login_num')->comment('尝试登录数');
-            $table->timestamp('lock_time')->comment('登录锁定时间');
+            $table->timestamp('lock_time')->nullable()->comment('登录锁定时间');
             $table->tinyInteger('is_enable')->comment('是否启用');
             $table->string('email', 64)->comment('电邮');
             $table->string('phone', 64)->comment('手机');
@@ -38,6 +41,9 @@ class CreateMembersTable extends Migration
      */
     public function down()
     {
+        if (!Schema::hasTable('members')) {
+            return;
+        }
         Schema::drop('members');
     }
 }
