+++ database/migrations/2016_08_18_063344_create_messages_table.php
@@ -12,14 +12,17 @@ class CreateMessagesTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('messages')) {
+            return;
+        }
         Schema::create('messages', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
 
             $table->integer('send_id')->unsigned()->comment('0为系统发送');
             $table->integer('receive_id')->unsigned()->comment('0为系统接收');
-            $table->timestamp('send_time')->comment('发送时间');
-            $table->timestamp('receive_time')->comment('查看时间');
+            $table->timestamp('send_time')->nullable()->comment('发送时间');
+            $table->timestamp('receive_time')->nullable()->comment('查看时间');
             $table->mediumText('content')->nullable()->comment('发送的内容');
         });
     }
@@ -31,6 +34,9 @@ class CreateMessagesTable extends Migration
      */
     public function down()
     {
+        if (!Schema::hasTable('messages')) {
+            return;
+        }
         Schema::drop('messages');
     }
 }
