+++ database/migrations/2016_08_18_063346_create_navigations_table.php
@@ -12,6 +12,9 @@ class CreateNavigationsTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('navigations')) {
+            return;
+        }
         Schema::create('navigations', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
@@ -30,6 +33,9 @@ class CreateNavigationsTable extends Migration
      */
     public function down()
     {
+        if (!Schema::hasTable('navigations')) {
+            return;
+        }
         Schema::drop('navigations');
     }
 }
