+++ database/migrations/2016_08_18_063346_create_quests_table.php
@@ -12,6 +12,9 @@ class CreateQuestsTable extends Migration
      */
     public function up()
     {
+        if (Schema::hasTable('quests')) {
+            return;
+        }
         Schema::create('quests', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
@@ -21,8 +24,8 @@ class CreateQuestsTable extends Migration
             $table->integer('max_portion')->unsigned()->comment('最大份数');
             $table->mediumText('start_content')->nullable()->comment('欢迎词');
             $table->mediumText('end_content')->nullable()->comment('结束词');
-            $table->timestamp('start_time')->comment('开始时间');
-            $table->timestamp('end_time')->comment('结束时间');
+            $table->timestamp('start_time')->nullable()->comment('开始时间');
+            $table->timestamp('end_time')->nullable()->comment('结束时间');
             $table->mediumText('access_info')->nullable()->comment('访问配置');
             $table->mediumText('ext_info')->nullable()->comment('后面说明');
         });
@@ -35,6 +38,9 @@ class CreateQuestsTable extends Migration
      */
     public function down()
     {
+        if (!Schema::hasTable('quests')) {
+            return;
+        }
         Schema::drop('quests');
     }
 }
